Add es-module build for easier tree-shaking
